requirements.txt
dwave-ocean-sdk==3.3.0
matplotlib==3.5.0
pandas==1.3.4
seaborn==0.11.2
sklearn
tables==3.6.1
tqdm
